feat: useAuth
