+++ src/pages/Signup.js
@@ -1,26 +1,14 @@
 import axios from "axios";
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import useAuth from "../hooks/useAuth";
 
 const BASE_URL = "https://www.pre-onboarding-selection-task.shop";
 
 function Signup() {
   const navigate = useNavigate();
-  const [disabled, setDisabled] = useState(true);
-  const [inputs, setInputs] = useState({
-    email: "",
-    password: "",
-  });
-
-  const { email, password } = inputs;
-
-  const onChange = (event) => {
-    const { value, name } = event.target;
-    setInputs({
-      ...inputs,
-      [name]: value,
-    });
-  };
+  const { email, password, disabled, onEmailChange, onPasswordChange } =
+    useAuth();
 
   const onSubmit = async (event) => {
     event.preventDefault();
@@ -50,12 +38,6 @@ function Signup() {
     }
   }, [navigate]);
 
-  useEffect(() => {
-    const isEmailValid = email.includes("@");
-    const isPasswordValid = password.length >= 8;
-    setDisabled(!(isEmailValid && isPasswordValid));
-  }, [email, password]);
-
   return (
     <div>
       <form onSubmit={onSubmit}>
@@ -64,14 +46,14 @@ function Signup() {
           data-testid="email-input"
           name="email"
           value={email}
-          onChange={onChange}
+          onChange={onEmailChange}
         />
         <input
           type="password"
           data-testid="password-input"
           name="password"
           value={password}
-          onChange={onChange}
+          onChange={onPasswordChange}
         />
         <button data-testid="signup-button" disabled={disabled}>
           회원가입
